cas: add cas_configuration symlink
Bug: T244849

+++ scap/checks/linkconf.sh
@@ -10,6 +10,11 @@ if [ ! -f ${DEPLOY_DIR}/src/netbox/netbox/configuration.py ]; then
 	ln -s /etc/netbox/configuration.py ${DEPLOY_DIR}/src/netbox/netbox/configuration.py
 fi
 
-if [ ! -f ${DEPLOY_DIR}/src/netbox/netbox/ldap_config.py ]; then
+
+if [ ! -f ${DEPLOY_DIR}/src/netbox/netbox/ldap_config.py ] && [ -f /etc/netbox/ldap.py ]; then
     ln -s /etc/netbox/ldap.py ${DEPLOY_DIR}/src/netbox/netbox/ldap_config.py
+fi
+
+if [ ! -f ${DEPLOY_DIR}/src/netbox/netbox/cas_configuration.py ] && [ -f /etc/netbox/cas_configuration.py ]; then
+    ln -s /etc/netbox/cas_configuration.py ${DEPLOY_DIR}/src/netbox/netbox/cas_configuration.py
 fi
